backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=5.0.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
pgvector>=0.2.4
sentence-transformers>=2.3.1
langchain-groq>=0.0.1
langchain-huggingface>=0.1.0
